requirements.txt
flask
requests==2.31.0
fuzzywuzzy
python-dotenv
openai==0.28
openpyxl
schedule
python-Levenshtein
gunicorn==20.1.0
google-cloud-bigquery==3.10.0
google-cloud-bigquery-storage==2.28.0
psutil==5.9.0
watchdog==3.0.0
decorator==5.1.1
numpy==1.21.6
pandas==1.3.5
db-dtypes == 1.4.1